+++ lib/country_meta.dart
@@ -12,4 +12,5 @@ export 'src/flags/flag_image_format.dart';
 export 'src/flags/flag_heigth.dart';
 export 'src/flags/flag_url_builder.dart';
 export 'src/flags/flag_width.dart';
+export 'src/flags/flag_emoji_builder.dart';
 export 'src/flags/country_flag.dart';